refactoring: create packages for feedback_screen, schedule_screen and ui package for common ui widgets
